Add information about weekly and monthly mirror
This adds an explanation, about which mirror is refreshed weekly and monthly. Since the original forum post is silent about which of both is weekly and which is monthly, I make a guess. 
The point is obviously, to prevent switching to that mirror for e.g performance related issues and finding the system in an old state and not knowing what is going on.